+++ jest-preprocess.js
@@ -2,7 +2,7 @@
 // Licensed under the MIT License
 
 const babelOptions = {
-  presets: ['babel-preset-gatsby'],
+  presets: ['babel-preset-gatsby', '@babel/preset-typescript'],
 };
 
 module.exports = require('babel-jest').createTransformer(babelOptions);